refactor: create dto and move files
