+++ apps/auth-api/src/application/users/repositories/events/dto/user-created-message-value.dto.ts
@@ -1,0 +1,9 @@
+import { IsEmail, IsString } from 'class-validator';
+
+export class UserCreatedMessageValue {
+  @IsString()
+  id!: string;
+
+  @IsEmail()
+  email!: string;
+}
