+++ apps/auth-api/src/application/users/repositories/events/topics/index.ts
