+++ apps/auth-api/src/application/users/users.service.ts
@@ -1,17 +1,17 @@
 import { Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
+import { UsersDatabaseRepository } from './repositories/database/users-database.repository';
 import { UsersEventsRepository } from './repositories/events/users-events.repository';
-import { UsersStorageRepository } from './repositories/storage/users-storage.repository';
 
 @Injectable()
 export class UsersService {
   constructor(
-    private usersStorageRepository: UsersStorageRepository,
+    private usersDatabaseRepository: UsersDatabaseRepository,
     private usersEventsRepository: UsersEventsRepository
   ) {}
 
   async create(createUserDto: CreateUserDto) {
-    const user = await this.usersStorageRepository.create(createUserDto);
+    const user = await this.usersDatabaseRepository.create(createUserDto);
     await this.usersEventsRepository.publishUserCreated(user);
     return `Hello, ${createUserDto.email}!`;
   }
